refactor: undo rename of sidebar component
